==> 03.Refine_Divide.ipynb <==
# Data division and purification

import pandas as pd
from ast import literal_eval

Since we gonna use only AlbumCover containing facial information

df0427 = pd.read_csv('./dataset_imsi/album_data_0522.csv')

facial = df0427.copy()
facial = facial[facial['age'] != 0]
facial = facial.reset_index()
facial = facial.drop('index', axis=1)

idx_nm = facial[(facial['Brass & Military'] == 1) | (facial["Children's"] == 1) | (facial['Classical'] == 1) | (facial['Non-Music'] == 1) | (facial['Stage & Screen'] == 1)| (facial['Blues'] == 1) | (facial['Latin'] == 1) | (facial['Reggae'] == 1)].index
facial = facial.drop(idx_nm)
facial = facial.drop(['Brass & Military', "Children's", 'Classical', 'Non-Music', 'Stage & Screen', 'Blues', 'Latin', 'Reggae'], axis=1)

if you want to shuffle the data, use the code below

# import sklearn

# imsi = sklearn.utils.shuffle(imsi)

Also we have to make test data

# Not now

Seperate to X and y

facial

facial_X = facial.iloc[:,:9]

facial_y = facial.iloc[:,9:]


facial_y

Make Y_label CSV file

facial_y.to_csv('./dataset/y_label.csv', index=False)

from X to IMG / Facial Info / Color Info

# Input Layer 1 : image (facial_x_img)
facial_x_img = facial_X[['path']]

# Input Layer 2 : facial info. (facial_x_face)
facial_x_face_imsi = facial_X[['age', 'gender', 'dominant_race', 'emotion']]
facial_x_face_imsi = facial_x_face_imsi.assign(angry=0.0, disgust=0.0, fear=0.0, happy=0.0, sad=0.0, surprise=0.0, neutral=0.0)

for i in facial_x_face_imsi.index:
    imsi_dict = literal_eval(facial_x_face_imsi['emotion'][i])
    
    for emo in imsi_dict:
        facial_x_face_imsi[emo][i] = imsi_dict[emo]
        
facial_x_face = facial_x_face_imsi.drop(['emotion'], axis=1)

# Input Layer 3 : dominant color. (facial_x_clr)
facial_x_clr = facial_X[['dominant_r', 'dominant_g','dominant_b']]

pd.DataFrame.to_csv(facial_x_face, './dataset/facial.csv', index=False)
pd.DataFrame.to_csv(facial_x_img, './dataset/img.csv', index=False)
pd.DataFrame.to_csv(facial_x_clr, './dataset/clr.csv', index=False)

# also save original dataframe to reuse
pd.DataFrame.to_csv(facial, './dataset/onlyface_with_y.csv', index=False)
